refactor: change package to Enceladus
