+++ src/main/kotlin/io/soulglo/enceladus/NucleusApplication.kt
@@ -1,4 +1,4 @@
-package io.soulglo.nucleus
+package io.soulglo.enceladus
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
