+++ src/test/kotlin/io/soulglo/enceladus/chapter/ComposeServiceTests.kt
@@ -1,4 +1,4 @@
-package io.soulglo.nucleus.chapter
+package io.soulglo.enceladus.chapter
 
 import org.jsoup.nodes.Document
 import org.jsoup.nodes.Element
